requirements.txt
streamlit
google-generativeai
openpyxl
